Make this a package and use automatic script creation
